feat: product, coupon 훅 통합
