+++ src/basic/entities/coupons/hooks/useCoupons.ts
@@ -1,9 +1,14 @@
 import { initialCoupons } from '@/basic/constants/mocks';
 import { useLocalStorage } from '@/basic/hooks';
 import { Coupon } from '@/types';
-import { useState } from 'react';
+import { useCallback, useState } from 'react';
+import { validateCouponCode } from '@/basic/utils';
 
-export function useCoupons() {
+interface UseCouponsProps {
+  addNotification: (message: string, type: 'error' | 'success' | 'warning') => void;
+}
+
+export function useCoupons({ addNotification }: UseCouponsProps) {
   const [coupons, setCoupons] = useLocalStorage<Coupon[]>('coupons', initialCoupons);
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
   const [showCouponForm, setShowCouponForm] = useState(false);
@@ -13,6 +18,30 @@ export function useCoupons() {
     discountType: 'amount' as 'amount' | 'percentage',
     discountValue: 0,
   });
+
+  const addCoupon = useCallback(
+    (newCoupon: Coupon) => {
+      const validation = validateCouponCode(newCoupon.code, coupons);
+      if (!validation.isValid) {
+        addNotification(validation.errorMessage!, 'error');
+        return;
+      }
+      setCoupons((prev) => [...prev, newCoupon]);
+      addNotification('쿠폰이 추가되었습니다.', 'success');
+    },
+    [coupons, addNotification, setCoupons]
+  );
+
+  const deleteCoupon = useCallback(
+    (couponCode: string) => {
+      setCoupons((prev) => prev.filter((c) => c.code !== couponCode));
+      if (selectedCoupon?.code === couponCode) {
+        setSelectedCoupon(null);
+      }
+      addNotification('쿠폰이 삭제되었습니다.', 'success');
+    },
+    [selectedCoupon, addNotification, setCoupons]
+  );
 
   return {
     coupons,
@@ -24,5 +53,8 @@ export function useCoupons() {
     setSelectedCoupon,
     setShowCouponForm,
     setCouponForm,
+
+    addCoupon,
+    deleteCoupon,
   };
 }
