+++ src/basic/entities/products/hooks/useProducts.ts
@@ -1,8 +1,12 @@
 import { initialProducts, ProductWithUI } from '@/basic/constants/mocks';
 import { useLocalStorage } from '@/basic/hooks';
-import { useState } from 'react';
+import { useCallback, useState } from 'react';
 
-export function useProducts() {
+interface UseProductsProps {
+  addNotification: (message: string, type: 'error' | 'success' | 'warning') => void;
+}
+
+export function useProducts({ addNotification }: UseProductsProps) {
   const [products, setProducts] = useLocalStorage<ProductWithUI[]>('products', initialProducts);
   const [editingProduct, setEditingProduct] = useState<string | null>(null);
   const [showProductForm, setShowProductForm] = useState(false);
@@ -13,6 +17,37 @@ export function useProducts() {
     description: '',
     discounts: [] as Array<{ quantity: number; rate: number }>,
   });
+
+  const addProduct = useCallback(
+    (newProduct: Omit<ProductWithUI, 'id'>) => {
+      const product: ProductWithUI = {
+        ...newProduct,
+        id: `p${Date.now()}`,
+      };
+      setProducts((prev) => [...prev, product]);
+      addNotification('상품이 추가되었습니다.', 'success');
+    },
+    [addNotification, setProducts]
+  );
+
+  const updateProduct = useCallback(
+    (productId: string, updates: Partial<ProductWithUI>) => {
+      setProducts((prev) =>
+        prev.map((product) => (product.id === productId ? { ...product, ...updates } : product))
+      );
+      addNotification('상품이 수정되었습니다.', 'success');
+    },
+    [addNotification, setProducts]
+  );
+
+  const deleteProduct = useCallback(
+    (productId: string) => {
+      setProducts((prev) => prev.filter((p) => p.id !== productId));
+      addNotification('상품이 삭제되었습니다.', 'success');
+    },
+    [addNotification, setProducts]
+  );
+
   return {
     products,
     setProducts,
@@ -22,5 +57,8 @@ export function useProducts() {
     setShowProductForm,
     productForm,
     setProductForm,
+    addProduct,
+    updateProduct,
+    deleteProduct,
   };
 }
